+++ test/fliff-get.spec.js
@@ -43,11 +43,46 @@ describe('fliff get', () => {
         const expectedResult = [
             { 'LIFF ID': '12345', 'Type': 'compact', 'URL': 'https://blah1', 'View': 'name2' },
             { 'LIFF ID': 'abcde', 'Type': 'full', 'URL': 'https://blah2', 'View': 'name1' }];
+        const options = {};
         let rsSend;
 
         beforeAll(async () => {
             jest.spyOn(request, 'send').mockResolvedValue(fakeSentResult);
-            rsSend = await fliff.get();
+            rsSend = await fliff.get(options);
+        });
+
+        afterAll(() => {
+            request.send.mockRestore();
+        });
+
+        it('sent', () => {
+            expect(request.send).toHaveBeenCalled();
+        });
+
+        it('return correct result', () => {
+            expect(rsSend).toEqual(expectedResult);
+        });
+
+    });
+
+    describe('when has detail option', () => {
+        const fakeSentResult = {
+            data: {
+                apps: [
+                    { liffId: '12345', view: { type: 'compact', url: 'https://blah1' }, description: 'Blah blah' },
+                    { liffId: 'abcde', view: { type: 'full', url: 'https://blah2' }, features: { ble: true } }
+                ]
+            }
+        };
+        const expectedResult = [
+            { 'LIFF ID': '12345', 'Type': 'compact', 'URL': 'https://blah1', 'View': 'name2', 'Description': 'Blah blah', 'BLE': '\u2613' },
+            { 'LIFF ID': 'abcde', 'Type': 'full', 'URL': 'https://blah2', 'View': 'name1', 'Description': '', 'BLE': '\u2713' }];
+        const options = { detail: true };
+        let rsSend;
+
+        beforeAll(async () => {
+            jest.spyOn(request, 'send').mockResolvedValue(fakeSentResult);
+            rsSend = await fliff.get(options);
         });
 
         afterAll(() => {